Add optional --producer-id concept
The Task that produces results sets --producer-id to its own name.

The Task that consumes results sets --producer-id to the name of the
task whose results its expects to consume.

An external system then verifies, based on the .spec.params and
.status.results of TaskRuns:

* No two tasks have a result with the same "producer_id"
* All params with a given "producer_id" match any result with the same
  "producer_id"

This prevents user modifications of the param values. If a task expects
to receive the result of another task, but the user passes something
else as the param, the external system catches this discrepancy.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>